feat: introduce base article component
